book/book/requirements.txt
django
gunicorn
django-allauth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
